Добавил удаление и редактирование 07.11.2021

+++ src/main/java/ru/university/app/university/repo/UsersRepo.java
@@ -6,6 +6,6 @@ import ru.university.app.university.models.UserUniversity;
 
 import java.util.Optional;
 
-public interface UsersRepo extends JpaRepository<UserUniversity, Integer> {
+public interface UsersRepo extends JpaRepository<UserUniversity, Long> {
     Optional<UserUniversity> findByEmail(String email);
 }
